refactor: jsp view model 처리 관련 로직 이전
