+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspView.java
@@ -1,5 +1,6 @@
 package webmvc.org.springframework.web.servlet.view;
 
+import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
@@ -12,7 +13,15 @@ public class JspView extends AbstractView {
 
     @Override
     protected void renderInternal(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        final var requestDispatcher = request.getRequestDispatcher(viewName);
+        addModelAttributes(model, request);
+        RequestDispatcher requestDispatcher = request.getRequestDispatcher(viewName);
         requestDispatcher.forward(request, response);
+    }
+
+    private void addModelAttributes(Map<String, ?> model, HttpServletRequest request) {
+        model.keySet().forEach(key -> {
+            log.debug("attribute name : {}, value : {}", key, model.get(key));
+            request.setAttribute(key, model.get(key));
+        });
     }
 }
